chore: moving test files into platform specific places
