+++ packages/star-button/__tests__/android/star-button.android.test.js
@@ -1,0 +1,3 @@
+import shared from "../shared.native";
+
+shared();
