--- packages/star-button/__tests__/android/star-button.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared";
-
-shared();
