+++ packages/star-button/__tests__/ios/star-button.ios.test.js
@@ -1,0 +1,3 @@
+import shared from "../shared.native";
+
+shared();
