--- packages/star-button/__tests__/ios/star-button.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared";
-
-shared();
