+++ packages/star-button/__tests__/shared.native.js
@@ -1,0 +1,23 @@
+import {
+  addSerializers,
+  compose,
+  flattenStyleTransform,
+  minimaliseTransform,
+  minimalNativeTransform,
+  print
+} from "@times-components/jest-serializer";
+import shared from "./shared.base";
+
+export default () => {
+  addSerializers(
+    expect,
+    compose(
+      print,
+      flattenStyleTransform,
+      minimalNativeTransform,
+      minimaliseTransform((value, key) => key === "style")
+    )
+  );
+
+  shared();
+};
