+++ packages/star-button/__tests__/shared.web.js
@@ -1,0 +1,22 @@
+import { AppRegistry } from "react-native-web";
+import {
+  addSerializers,
+  compose,
+  minimalWebTransform,
+  rnwTransform,
+  stylePrinter
+} from "@times-components/jest-serializer";
+import shared from "./shared.base";
+
+export default () => {
+  addSerializers(
+    expect,
+    compose(
+      stylePrinter,
+      minimalWebTransform,
+      rnwTransform(AppRegistry)
+    )
+  );
+
+  shared();
+};
