+++ packages/star-button/__tests__/web/star-button.web.test.js
@@ -1,0 +1,3 @@
+import shared from "../shared.web";
+
+shared();
