+++ packages/star-button/package.json
@@ -38,6 +38,8 @@
   "devDependencies": {
     "@thetimes/jest-lint": "*",
     "@times-components/eslint-config-thetimes": "0.8.9",
+    "@times-components/jest-serializer": "3.2.0",
+    "@times-components/test-utils": "2.2.14",
     "@times-components/jest-configurator": "2.3.4",
     "@times-components/storybook": "3.4.5",
     "@times-components/webpack-configurator": "2.0.15",
@@ -53,7 +55,7 @@
     "webpack-cli": "2.1.4"
   },
   "dependencies": {
-    "@times-components/icons": "2.8.0",
+    "@times-components/icons": "2.8.1",
     "@times-components/link": "3.2.30",
     "@times-components/styleguide": "3.26.3",
     "prop-types": "15.6.2"
